fix warnigns on windows

+++ src/tables/column.hpp
@@ -38,11 +38,11 @@ namespace tables
 
         /// Returns the number of rows
         /// @return The number of rows
-        virtual uint32_t rows() const = 0;
+        virtual size_t rows() const = 0;
 
         /// Returns the number of empty rows
         /// @return The number of empty rows
-        virtual uint32_t empty_rows() const = 0;
+        virtual uint64_t empty_rows() const = 0;
 
         /// Sets a value in the column, at the current row
         /// @param value The value to set
