+++ src/tables/nonconst_column.hpp
@@ -38,10 +38,10 @@ namespace tables
         void add_row();
 
         /// @copydoc column::rows() const
-        uint32_t rows() const;
+        size_t rows() const;
 
         /// @copydoc column::empty_rows() const
-        uint32_t empty_rows() const;
+        uint64_t empty_rows() const;
 
         /// @copydoc column::set_value(const boost::any&)
         void set_value(const boost::any& value);
